Fixed Cygwin tmp dir symlink problem. As was written, when splice restarted, contents of /cygdrive/c/tmp were copied into /tmp_bak, as if when run the first time. The fix was to recognize that /tmp had already been symlinked to /cygdrive/c/tmp.

+++ structured_derby/src/main/resources/bin/start-splice.sh
@@ -20,10 +20,10 @@ CYGWIN=`uname -s`
 if [[ ${CYGWIN} == CYGWIN* ]]; then
     # cygwin likes to write in 2 places for /tmp
     # we'll symlink them
-    if [[ -e "/tmp" ]]; then
+    if [[ -e "/tmp" && ! -L "/tmp" ]]; then
         mv "/tmp" "/tmp_bak"
+        ln -s "/cygdrive/c/tmp" "/tmp"
     fi
-    ln -s "/cygdrive/c/tmp" "/tmp"
 fi
 # start zookeeper once
 "${ROOT_DIR}"/bin/_startZoo.sh "${ROOT_DIR}" "${LOGFILE}" "${DEBUG}"
